test(actionpack): relocate mount tests to dispatch/mapper.test.ts for test:compare
Rails' mapper_test.rb lives at actionpack/test/dispatch/, which by
convention maps to packages/actionpack/src/action-dispatch/dispatch/
mapper.test.ts. Co-locating the ported mount tests there (alongside the
existing 'initialize' test under describe('MapperTest', ...)) flips them
from 'misplaced' to 'matched' in test:compare. Renamed to the Rails
description form ('can pass anchor to mount' etc.) since the test:compare
extractor strips the 'test_' prefix and converts underscores to spaces.
